Also set DISABLE_MANGOHUD and fix unsetenv usage.
DISABLE_MANGOHUD=1 and unset MANGOHUD,
to ask Vulkan Loader to not even load the mangohud layer.

From my tests if both DISABLE_MANGOHUD=1 and MANGOHUD=1,
is set, the layer is NOT loaded, but I can't find this documented,
so be careful and set one and unset another.

`getenv` returns a pointer that is not under app control.

If `unsetenv` is called, then this pointer can't be reused
for later `setenv`.

See IEEE Std 1003.1-2017 ("POSIX 2017"), or previous editions.

https://pubs.opengroup.org/onlinepubs/9699919799/functions/getenv.html

Most importantly this section:

> The returned string pointer might be invalidated or  the string content might be overwritten by a subsequent call to getenv(),  setenv(), unsetenv(), ...

Restoring of the `LD_PRELOAD` at line 704, never probably worked,
but also probably never got to the point to mater. Fix intention.

Additionally `setenv` should not have second argument be NULL,
as this is undefined behaviour according to POSIX.
Some older versions of glibc supported that to mean unset,
but now they return EINVAL instead. Some other C libraries
(notably IBM and QNX) do accept NULL (being equivalent to
calling `unsetenv`), but that is not behaviour guaranteed by
standard. This is why there is separate path for `setenv`
and `unsetenv` for `DISABLE_MANGOHUD`.