rpc: Split receive into smaller functions. Deindent / fix indents. Add comments.
This hopefully makes it a bit easier to follow.

Add comments around.

Invert some conditions, so the code can be deindentet a bit,
which should help with readability too. Fix minor indent / code
align mistakes.

Add few more asserts here and there, more as a documentation
of the logic, and to help protect mistakes in future refactors.